pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "carm"
version = "0.0.20250822"
authors = [
    {name = "Yong Zhao", email = "zhaoyong11933@cvte.com"},
]
description = "Python interface for cvte arm."
requires-python = ">=3.6"
dependencies = ["websocket-client"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

[project.scripts]
carm = "carm.carm_dds:main"
carm_ros2 = "carm.carm_ros2:main"

[tool.setuptools]
packages = ["carm"]

[tool.setuptools.package-data]
carm = [
    "__init__.py",
    "carm.py",
    "carm_dds.py",
]
